Do actions, add ajaxurl, and add div#widgets-right for compat
Widget form controls expect to be set up on the widgets page, so the
actions triggered on that page need to be manually triggered here
(just like the Ajax handler does)

The widgets page also exposes an ajaxurl global JavaScript variable
which the Customizer does not export, so we have to set it as well.

Lastly, many widgets expect for their controls to be rendered inside
of a div#widgets-right which does not exist in the customizer, so we
have to make one of the divs in the customizer have that ID for
compatability.

See #39