Fade widgets not rendered in sidebars on currently-previewed URL
A 50% opacity is applied to the widget form controls when the widget
is not rendered in the preview, for example if the Jetpack Widget
Visibility module is used to exclude the widget from the page.

When navigating to other URLs in the preview, the appearance of the
widget controls will update based on whether or not the widget is
rendered on the newly-previewed URL. As such, this is the same
behavior as the sidebar sections.

Also fixed bug in relation to Widget Visibility where widgets not
initially-rendered in the preview could not then subsequently be
updated or previewed.

Fixes #65